Description added to the workshop
It's really important to give feedback about the training.
Especially considering the fact that I was so lucky to get
to this workshop within 1 minute after tickets were available.
I like to spend my saturday doing git stuff with those trainers.
But it's also mind-twisting how powerful and complicated is git.

Things still to learn:
[ ] cherry-pick
[ ] reset
[ ] revert